+++ packages/terra-menu/src/MenuItem.jsx
@@ -14,7 +14,24 @@ const cx = classNamesBind.bind(styles);
 
 const contextTypes = {
   isGroupItem: PropTypes.bool,
+
+  /**
+   * isSelectableMenu has been deprecated and will be removed on next major version release.
+   * Rename the `isSelectableMenu` prop to `isToggleableMenu`.
+   */
   isSelectableMenu: PropTypes.bool,
+
+  /**
+  * Indicates if the menu should be toggleable.
+  */
+  isToggleableMenu: PropTypes.bool,
+
+  /**
+   * Indicates if the menu needs to reserve space on the left for each menu item if one
+   * or more of the items needs to display a custom icon, instructions for use icon, or
+   * a checkmark if the item is part of a group item.
+   */
+  shouldReserveSpaceForIcon: PropTypes.bool,
 };
 
 const propTypes = {
@@ -29,14 +46,28 @@ const propTypes = {
   isDisabled: PropTypes.bool,
 
   /**
-   * Indicates if the item is selected. A selected item is indicated with a checkmark.
+   * ![IMPORTANT](https://badgen.net/badge/prop/deprecated/red)
+   * isSelected has been deprecated and will be removed on next major version release.
+   * Rename the `isSelected` prop to `isToggled`.
    */
   isSelected: PropTypes.bool,
 
   /**
-   * Indicates if the item should be selectable.
+   * ![IMPORTANT](https://badgen.net/badge/prop/deprecated/red)
+   * isSelectable has been deprecated and will be removed on next major version release.
+   * Rename the `isSelectable` prop to `isToggleable`.
    */
   isSelectable: PropTypes.bool,
+
+  /**
+  * Indicates if the item is toggled. A toggled item is indicated with a checkmark.
+  */
+  isToggled: PropTypes.bool,
+
+  /**
+  * Indicates if the item should be toggleable.
+  */
+  isToggleable: PropTypes.bool,
 
   /**
    * Displays the  eIFU (electronic instructions for use) icon for menu item if set to true. (This icon is used to indicate Help content that is the equivalent of an instruction manual)
@@ -44,7 +75,7 @@ const propTypes = {
   isInstructionsForUse: PropTypes.bool,
 
   /**
-   * List of Menu.Items to display in a submenu when this item is selected.
+   * List of Menu.Items to display in a submenu when this item is clicked.
    */
   subMenuItems: PropTypes.arrayOf(PropTypes.element),
 
@@ -54,7 +85,7 @@ const propTypes = {
   onClick: PropTypes.func,
 
   /**
-   * Callback function for when selection state changes on a selectable item.
+   * Callback function for when toggleable state changes on a toggleabe item.
    */
   onChange: PropTypes.func,
 
@@ -62,28 +93,51 @@ const propTypes = {
    * Indicates if the item has focus. This is used internally to control focus and does not set initial focus.
    */
   isActive: PropTypes.bool,
+
+  /**
+   * Custom icon to display in the item
+   */
+  icon: PropTypes.element,
+
+  /**
+   * @private
+   * Indicates if the item should display with a highlighted background. Reserved for Terra higher-order component approved usage only.
+   */
+  isHighlighted: PropTypes.bool,
 };
 
 const defaultProps = {
   text: '',
-  isSelected: false,
+  isToggled: false,
   isInstructionsForUse: false,
   isActive: false,
-  isSelectable: undefined,
+  isToggleable: undefined,
   isDisabled: false,
   subMenuItems: [],
+  isHighlighted: false,
 };
 
+// TODO: remove isSelect and isSeletable props on the next major release
 class MenuItem extends React.Component {
   constructor(props, context) {
     super(props, context);
+    const {
+      isSelected,
+      isSelectable,
+      isToggled,
+      isToggleable,
+    } = this.props;
+
     this.wrapOnClick = this.wrapOnClick.bind(this);
     this.wrapOnKeyDown = this.wrapOnKeyDown.bind(this);
     this.wrapOnKeyUp = this.wrapOnKeyUp.bind(this);
-    this.handleSelection = this.handleSelection.bind(this);
+    this.handleToggled = this.handleToggled.bind(this);
     this.setItemNode = this.setItemNode.bind(this);
+    const toggled = (isToggled || isSelected);
+    const toggleable = (isToggleable || isSelectable);
+
     this.state = {
-      isSelected: props.isSelected && props.isSelectable && !context.isGroupItem,
+      isToggled: toggled && toggleable && !context.isGroupItem,
       isActive: false,
     };
   }
@@ -94,14 +148,19 @@ class MenuItem extends React.Component {
     }
   }
 
-  handleSelection(event) {
+  handleToggled(event) {
     event.preventDefault();
 
-    if (this.props.isSelectable && !this.context.isGroupItem && !this.props.isDisabled) {
-      this.setState(prevState => ({ isSelected: !prevState.isSelected }));
+    const {
+      isSelectable,
+      isToggleable,
+    } = this.props;
+
+    if ((isToggleable || isSelectable) && !this.context.isGroupItem && !this.props.isDisabled) {
+      this.setState(prevState => ({ isToggled: !prevState.isToggled }));
 
       if (this.props.onChange) {
-        this.props.onChange(event, !this.state.isSelected);
+        this.props.onChange(event, !this.state.isToggled);
       }
     }
   }
@@ -113,7 +172,7 @@ class MenuItem extends React.Component {
   }
 
   wrapOnClick(event) {
-    this.handleSelection(event);
+    this.handleToggled(event);
 
     if (this.props.onClick) {
       this.props.onClick(event);
@@ -123,7 +182,7 @@ class MenuItem extends React.Component {
   wrapOnKeyDown(onKeyDown) {
     return ((event) => {
       if (event.nativeEvent.keyCode === KeyCode.KEY_RETURN || event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
-        this.handleSelection(event);
+        this.handleToggled(event);
 
         // Only add active style if the menu doesn't have a sub menu to avoid active style being stuck enabled
         if (!(this.props.subMenuItems && this.props.subMenuItems.length > 0)) {
@@ -161,18 +220,31 @@ class MenuItem extends React.Component {
       text,
       isDisabled,
       isSelected,
+      isToggled,
       isInstructionsForUse,
       isSelectable,
+      isToggleable,
       subMenuItems,
       isActive,
+      icon,
+      isHighlighted,
       ...customProps
     } = this.props;
 
-    const { isGroupItem, isSelectableMenu } = this.context;
-
+    const {
+      isGroupItem,
+      isToggleableMenu,
+      isSelectableMenu,
+      shouldReserveSpaceForIcon,
+    } = this.context;
     const attributes = { ...customProps };
+
     attributes.tabIndex = isDisabled ? '-1' : '0';
     attributes['aria-disabled'] = isDisabled;
+
+    const toggled = (isToggled || isSelected);
+    const toggleable = (isToggleable || isSelectable);
+    const toggleableMenu = (isToggleableMenu || isSelectableMenu);
 
     // This is passed down by the single select list in group item and not needed
     delete attributes.hasChevron;
@@ -186,13 +258,15 @@ class MenuItem extends React.Component {
       attributes.onKeyUp = this.wrapOnKeyUp(attributes.Up);
     }
 
-    const markAsSelected = this.state.isSelected || (isGroupItem && isSelected);
+    const markAsToggled = this.state.isToggled || (isGroupItem && toggled);
     const textContainer = <div className={cx('text')}>{text}</div>;
     const hasChevron = subMenuItems.length > 0;
 
     const itemClassNames = cx([
       'item',
-      { selected: markAsSelected },
+      { 'is-highlighted': isHighlighted },
+      { 'is-toggled': markAsToggled },
+      { 'is-toggleable': toggleable },
       { 'is-disabled': isDisabled },
       { 'is-top-level': hasChevron },
       // eslint-disable-next-line quote-props
@@ -201,11 +275,19 @@ class MenuItem extends React.Component {
     ]);
 
     let content = textContainer;
-    if (hasChevron || isSelectableMenu || isInstructionsForUse) {
+    if (hasChevron || toggleableMenu || isInstructionsForUse || icon || shouldReserveSpaceForIcon) {
       let fitStartIcon = null;
       if (isInstructionsForUse) {
         fitStartIcon = <InstructionsForUseIcon className={cx('start-icon')} />;
-      } else if (isSelectableMenu) {
+      } else if (toggleableMenu) {
+        if (!toggleable && icon) {
+          fitStartIcon = React.cloneElement(icon, { className: cx('start-icon') });
+        } else {
+          fitStartIcon = <CheckIcon className={cx(['checkmark', 'start-icon'])} />;
+        }
+      } else if (icon) {
+        fitStartIcon = React.cloneElement(icon, { className: cx('start-icon') });
+      } else if (shouldReserveSpaceForIcon) {
         fitStartIcon = <CheckIcon className={cx(['checkmark', 'start-icon'])} />;
       }
 
@@ -222,14 +304,14 @@ class MenuItem extends React.Component {
     let role = 'menuitem';
     if (isGroupItem) {
       role = 'menuitemradio';
-    } else if (isSelectable) {
+    } else if (toggleable) {
       role = 'menuitemcheckbox';
     }
 
     return (
       <ThemeContext.Consumer>
         { theme => (
-          <li {...attributes} className={classNames(itemClassNames, cx(theme.className))} ref={this.setItemNode} role={role} aria-checked={markAsSelected}>
+          <li {...attributes} className={classNames(itemClassNames, cx(theme.className))} ref={this.setItemNode} role={role} aria-checked={markAsToggled}>
             {content}
           </li>
         )}