+++ packages/terra-menu/src/_MenuContent.jsx
@@ -11,6 +11,7 @@ import classNames from 'classnames/bind';
 import * as KeyCode from 'keycode-js';
 import ThemeContext from 'terra-theme-context';
 import MenuUtils from './_MenuUtils';
+import MenuItem from './MenuItem';
 import styles from './Menu.module.scss';
 
 const cx = classNames.bind(styles);
@@ -97,7 +98,8 @@ const defaultProps = {
 };
 
 const childContextTypes = {
-  isSelectableMenu: PropTypes.bool,
+  isToggleableMenu: PropTypes.bool,
+  shouldReserveSpaceForIcon: PropTypes.bool,
 };
 
 class MenuContent extends React.Component {
@@ -106,7 +108,8 @@ class MenuContent extends React.Component {
     this.wrapOnClick = this.wrapOnClick.bind(this);
     this.wrapOnKeyDown = this.wrapOnKeyDown.bind(this);
     this.buildHeader = this.buildHeader.bind(this);
-    this.isSelectable = this.isSelectable.bind(this);
+    this.isToggleable = this.isToggleable.bind(this);
+    this.shouldReserveSpaceForIcon = this.shouldReserveSpaceForIcon.bind(this);
     this.onKeyDown = this.onKeyDown.bind(this);
     this.onKeyDownBackButton = this.onKeyDownBackButton.bind(this);
     this.validateFocus = this.validateFocus.bind(this);
@@ -119,7 +122,7 @@ class MenuContent extends React.Component {
   }
 
   getChildContext() {
-    return { isSelectableMenu: this.isSelectable() };
+    return { isToggleableMenu: this.isToggleable(), shouldReserveSpaceForIcon: this.shouldReserveSpaceForIcon() };
   }
 
   componentDidUpdate(prevProps) {
@@ -174,15 +177,45 @@ class MenuContent extends React.Component {
     }
   }
 
-  isSelectable() {
-    let isSelectable = false;
+  isToggleable() {
+    let isToggleableValue = false;
     React.Children.forEach(this.props.children, (child) => {
-      if (child.props.children || child.props.isSelectable) {
-        isSelectable = true;
+      const {
+        children,
+        isToggleable,
+        ...customProps
+      } = child.props;
+
+      // child is a group menu item that needs space reserved for the checkmark
+      React.Children.forEach(children, (subchild) => {
+        if (subchild.type === MenuItem) {
+          isToggleableValue = true;
+        }
+      });
+
+      if (isToggleable || customProps.isSelectable) {
+        isToggleableValue = true;
       }
     });
 
-    return isSelectable;
+    return isToggleableValue;
+  }
+
+  shouldReserveSpaceForIcon() {
+    let shouldReserveSpaceForIcon = false;
+    React.Children.forEach(this.props.children, (child) => {
+      const {
+        icon,
+        isInstructionsForUse,
+      } = child.props;
+
+      // reserve space for when there is a custom icon or instructions icon to be shown
+      if (icon || isInstructionsForUse) {
+        shouldReserveSpaceForIcon = true;
+      }
+    });
+
+    return shouldReserveSpaceForIcon;
   }
 
   wrapOnClick(item) {
@@ -294,6 +327,7 @@ class MenuContent extends React.Component {
     let index = -1;
     let shouldDisplayMainMenuHeader;
     const items = this.props.children ? [] : undefined;
+
     React.Children.map(this.props.children, (item) => {
       const onClick = this.wrapOnClick(item);
       let newItem = item;