+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated jest snapshots
+
 ## 6.61.0 - (March 15, 2022)
 
 * Changed